\name{CondTimeROC-package}
\alias{CondTimeROC-package}
\alias{CondTimeROC}
\docType{package}
\title{
\packageTitle{CondTimeROC}
}
\description{
\packageDescription{CondTimeROC}
}
%\details{
%\packageDESCRIPTION{CondTimeROC}
%\packageIndices{CondTimeROC}
%}
\author{
\packageAuthor{CondTimeROC}

Maintainer: \packageMaintainer{CondTimeROC}
}
\references{

M.X Rodriguez-Alvarez and V. Inacio (2025). Penalised spline estimation of covariate-specific time-dependent ROC curves. Technical Report

M.X Rodriguez-Alvarez, L. Meira-Machado, E. Abu-Assi and S. Raposeiras-Roubin (2016). Nonparametric estimation of time-dependent ROC curves conditional on a continuous covariate. \emph{Statistics in Medicine} \bold{35}, 1090--1102.

X. Song and X-H. Zhou (2008). A semiparametric approach for the covariate specific ROC curve with survival outcome. \emph{Statistica Sinica} \bold{18}, 947--965.
}
\keyword{package}
%\seealso{}
%\examples{
%~~ simple examples of the most important functions ~~
%}
